feat: add js language server
add js language server in languageadapter,
could realize js code highlight、rename and
other features.

Log: js hightlight supported.

+++ src/plugins/codeeditor/mainframe/texteditkeeper.cpp
@@ -80,7 +80,7 @@ QString TextEditKeeper::getTokenTypeAnnLine(const QString &tokenType, const QStr
         if (tokenType.toStdString() != tokenMap.semanticTokenType
                 || tokenMap.result.empty())
             continue;
-        for (int idx = 0; idx < rules.size(); idx++) {
+        for (size_t idx = 0; idx < rules.size(); idx++) {
             if (!result.isEmpty()) result += " ";
             result += QString::fromStdString(data.rules[idx]);
             if (!result.isEmpty()) result += ": ";
